backend/requirements.txt
absl-py==2.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==5.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.2.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
deep-translator==1.11.4
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.45
groq==0.32.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langdetect==1.0.9
libretranslatepy==2.1.1
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mediapipe==0.10.21
ml_dtypes==0.5.1
mysql-connector-python==9.4.0
narwhals==2.7.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.12.0
pydantic_core==2.41.1
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.1
PyMySQL==1.1.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
scikit-learn==1.7.0
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
translate==3.6.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
